adjusted frontend to new game model logic
