+++ src/app/components/lobby-game/lobby-game.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { SocketKey } from 'src/app/models/object/service.model';
-import { Item, ItemAny, ItemType } from 'src/app/models/data/item.model';
+import { DeckItem, Item, ItemAny, ItemType } from 'src/app/models/data/item.model';
 import { TokenService } from 'src/app/services/auth/token.service';
 import { GameService } from 'src/app/services/game.service';
-import { SocketService } from 'src/app/services/request/socket.service';
 import { _Object } from 'src/app/models/data/object.model';
 import { DataService } from 'src/app/services/data.service';
 import { InventoryComponent } from '../inventory/inventory.component';
+import { RoomService } from 'src/app/services/room.service';
 
 @Component({
   selector: 'app-lobby-game',
@@ -18,6 +18,7 @@ export class LobbyGameComponent {
   @ViewChild('deckInventory') deckInventoryRef!: InventoryComponent;
 
   time: Date = new Date();
+  decks: _Object<DeckItem>[] = [];
 
   _selectedObject?: _Object<ItemAny>;
   set selectedObject(value: _Object<ItemAny> | undefined) {
@@ -28,7 +29,11 @@ export class LobbyGameComponent {
     return this._selectedObject;
   }
 
-  constructor(public game: GameService, public socket: SocketService, public data: DataService) { }
+  constructor(public game: GameService, public socket: RoomService, public data: DataService) { }
+
+  async loadDecks() {
+    this.decks = (await this.data.self.inventory).objects?.filter(x => ItemType.DECK == x.item.type) as _Object<DeckItem>[] ?? [];
+  }
 
   get SocketKey() {
     return SocketKey;
@@ -39,7 +44,7 @@ export class LobbyGameComponent {
   }
 
   get events(): { timestamp: Date; room: string; event: any }[] {
-    return this.socket.events.reverse().filter(x => x.timestamp > this.time);
+    return []; //this.socket.events.reverse().filter(x => x.timestamp > this.time);
   }
 
   async loadDeck(objectHash?: string) {
