+++ src/app/components/lobby-room/lobby-room.component.html
@@ -1,6 +1,6 @@
 <div class="container">
   <div class="header">
-    <button mat-icon-button (click)="socket.leave(SocketKey.ROOM)">
+    <button mat-icon-button (click)="room.leave(SocketKey.ROOM)">
       <mat-icon>arrow_back</mat-icon>
     </button>
     <h3>{{ name }}</h3>
@@ -14,7 +14,7 @@
   </div>
   <div class="actions">
     <span class="spacer"></span>
-    <button mat-flat-button color="primary" (click)="socket.join(SocketKey.GAME, name)">
+    <button mat-flat-button color="primary" (click)="room.join(SocketKey.GAME, name)">
       Join Game
     </button>
   </div>
