+++ src/app/components/lobby-room/lobby-room.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { SocketKey } from 'src/app/models/object/service.model';
-import { SocketService } from 'src/app/services/request/socket.service';
+import { RoomService } from 'src/app/services/room.service';
 
 @Component({
   selector: 'app-lobby-room',
@@ -11,7 +11,7 @@ export class LobbyRoomComponent {
 
   @Input() name?: string;
 
-  constructor(public socket: SocketService) { }
+  constructor(public room: RoomService) { }
 
   get SocketKey() {
     return SocketKey;
